tests for menuservicec addded

+++ src/main/java/de/iav/burgershop/repository/MenuRepository.java
@@ -1,5 +1,6 @@
 package de.iav.burgershop.repository;
 
+import de.iav.burgershop.exception.MenuNotFoundException;
 import de.iav.burgershop.model.Menu;
 import org.springframework.stereotype.Repository;
 
@@ -9,7 +10,7 @@ import java.util.Map;
 @Repository
 public class MenuRepository {
 
-    private Map<String, Menu> menuRepository;
+    private final Map<String, Menu> menuRepository;
 
     public MenuRepository(Map<String, Menu> menuRepository) {
         this.menuRepository = menuRepository;
@@ -20,18 +21,24 @@ public class MenuRepository {
     }
 
     public Menu getMenuByID(String id){
-        return this.menuRepository.get(id);
+        if (this.menuRepository.containsKey(id)){
+            return this.menuRepository.get(id);
+        }else throw new MenuNotFoundException(id);
     }
 
     public List<Menu> list(){
         return new ArrayList<Menu>(this.menuRepository.values());
     }
 
-    public void changeMenuByID(String menuIDTOChange, Menu menu) {
-        this.menuRepository.replace(menuIDTOChange,menu);
+    public void changeMenuByID(String menuIDToChange, Menu menu) {
+        if (this.menuRepository.containsKey(menuIDToChange)){
+            this.menuRepository.replace(menuIDToChange,menu);
+        }else throw new MenuNotFoundException(menuIDToChange);
     }
 
     public void deleteMenuByID(String menuIDToDelete) {
-        this.menuRepository.remove(menuIDToDelete);
+        if (this.menuRepository.containsKey(menuIDToDelete)){
+            this.menuRepository.remove(menuIDToDelete);
+        }else throw new MenuNotFoundException(menuIDToDelete);
     }
 }
